Added a compatibility layer
And fix bug with text version 2.0
by adding upper bound of the package.
