+++ Data/Double/Conversion/ByteString.hs
@@ -1,5 +1,3 @@
-{-# LANGUAGE TypeFamilies #-}
-
 -- |
 -- Module      : Data.Double.Conversion.ByteString
 -- Copyright   : (c) 2011 MailRank, Inc.
@@ -9,6 +7,10 @@
 -- Stability   : experimental
 -- Portability : GHC
 --
+-- This module left now only for compatibility and should not be used
+-- in new projects. 
+-- Please, use Convertable type class from Data.Double.Conversion.Convertable
+-- 
 -- Fast, efficient support for converting between double precision
 -- floating point values and text.
 --
@@ -18,27 +20,34 @@
 -- 'ByteString' values via @malloc@.)
 
 module Data.Double.Conversion.ByteString
-    ( convert
+    (
+      toExponential
+    , toFixed
+    , toPrecision
+    , toShortest
     ) where
 
-import Control.Monad (when)
-import Data.ByteString.Internal (ByteString(..), mallocByteString)
-import Data.Double.Conversion.FFI (ForeignFloating)
-import Data.Word (Word8)
-import Foreign.C.Types (CDouble, CFloat, CInt)
-import Foreign.ForeignPtr (withForeignPtr)
-import Foreign.Ptr (Ptr)
-import System.IO.Unsafe (unsafePerformIO)
+import qualified Data.Double.Conversion.Convertable
+import Data.ByteString.Internal
 
-convert :: (RealFloat a, RealFloat b , b ~ ForeignFloating a) => String -> CInt -> (b -> Ptr Word8 -> IO CInt)
-        -> a -> ByteString
-{-# SPECIALIZE convert :: String -> CInt -> (CDouble -> Ptr Word8 -> IO CInt) -> Double -> ByteString #-}
-{-# SPECIALIZE convert :: String -> CInt -> (CFloat -> Ptr Word8 -> IO CInt) -> Float -> ByteString #-}
-{-# INLINABLE convert #-}
-convert func len act val = unsafePerformIO $ do
-  fp <- mallocByteString (fromIntegral len)
-  size <- withForeignPtr fp $ act (realToFrac val)
-  when (size == -1) .
-    fail $ "Data.Double.Conversion.ByteString." ++ func ++
-           ": conversion failed (invalid precision requested)"
-  return $ PS fp 0 (fromIntegral size)
+-- | Compute a representation in exponential format with the requested
+-- number of digits after the decimal point. The last emitted digit is
+-- rounded.  If -1 digits are requested, then the shortest exponential
+-- representation is computed.
+toExponential :: Int -> Double -> ByteString
+toExponential = Data.Double.Conversion.Convertable.toExponential
+
+-- | Compute a decimal representation with a fixed number of digits
+-- after the decimal point. The last emitted digit is rounded.
+toFixed :: Int -> Double -> ByteString
+toFixed = Data.Double.Conversion.Convertable.toFixed
+
+-- | Compute the shortest string of digits that correctly represent
+-- the input number.
+toShortest :: Double -> ByteString
+toShortest = Data.Double.Conversion.Convertable.toShortest
+
+-- | Compute @precision@ leading digits of the given value either in
+-- exponential or decimal format. The last computed digit is rounded.
+toPrecision :: Int -> Double -> ByteString
+toPrecision = Data.Double.Conversion.Convertable.toPrecision
