+++ README.markdown
@@ -11,17 +11,11 @@ We are happy to receive bug reports, fixes, documentation enhancements,
 and other improvements.
 
 Please report bugs via the
-[github issue tracker](https://github.com/bos/double-conversion/issues).
+[github issue tracker](https://github.com/haskell/double-conversion/issues).
 
-Master [git repository](https://github.com/bos/double-conversion):
+Master [git repository](https://github.com/haskell/double-conversion):
 
-* `git clone git://github.com/bos/double-conversion.git`
-
-There's also a [Mercurial mirror](https://bitbucket.org/bos/double-conversion):
-
-* `hg clone https://bitbucket.org/bos/double-conversion`
-
-(You can create and contribute changes using either git or Mercurial.)
+* `git clone git://github.com/haskell/double-conversion`
 
 Authors
 -------
